// seductx/src/app/model/booking.model.ts
export interface Booking {
  booking_id: string;
  user_id: string;
  escort_id: string;
  hour: string;
  shot: string;
  price: string;
  user_name: string;
  user_email: string;
  special_request: string;
  hotel: string;
  room_no: string;
  booking_date: string;
  booking_end_date: string;
  booking_expiry_date: string;
  booking_location: string;
  booking_latitude: string;
  booking_longitude: string;
  total_amount: string;
  request_status: string;
  created_at: string;
  hour_name: string;
  hour_time: string;
  shot_name: string;
  user_first_name: string;
  user_last_name: string;
  user_country_code: string;
  user_phone_number: string;
  escort_first_name: string;
  escort_last_name: string;
  escort_latitude: string;
  escort_longitude: string;
  escort_email: string;
  escort_profile_image: string;
  escort_country_code: string;
  escort_phone_number: string;
  services: any;
  booking_time_hour: string;
  escort_rating: string;
  user_rating: string;
  request_status_name: string;
  time: string;
}